Fixed ammunition to be its own pbo
