--- addons/fix_tf47_m3maaws/CfgMagazines.hpp
@@ -1,20 +1,0 @@
-class CfgMagazines {
-    class CA_LauncherMagazine;
-    class tf47_m3maaws_HEAT_CS : CA_LauncherMagazine {
-        author="[TaskForce 47] BlackPixxel";
-        scope=2;
-        displayName="HEAT 665 CS";
-        model="tf47_launchers\m3maaws\mag_m3maaws.p3d";
-        picture="\tf47_launchers\m3maaws\UI\gear_small_CA.paa";
-        initSpeed=290;
-        descriptionShort="High Explosive Anti tank Confined Space";
-        displaynameshort="665 CS HEAT";
-        ammo="tf47_m3maaws_HEAT_CS";
-        type="2* 256";
-        mass=72.769997;
-        ace_overpressure_priority=99;
-        ace_overpressure_angle=50;
-        ace_overpressure_range=2;
-        ace_overpressure_damage=0.69999999;
-    };
-};
