+++ addons/weapons/$PBOPREFIX$
@@ -1,0 +1,1 @@
+z\cav\addons\weapons
