+++ addons/weapons/CfgMagazines.hpp
@@ -1,0 +1,20 @@
+class CfgMagazines {
+    class CA_LauncherMagazine;
+    class Cav_m3maaws_HEAT_CS : CA_LauncherMagazine {
+        author = "Grey.V";
+        scope = 2;
+        displayName = CSTRING(HEAT_655_CS);
+        model = "tf47_launchers\m3maaws\mag_m3maaws.p3d";
+        picture = "\tf47_launchers\m3maaws\UI\gear_small_CA.paa";
+        initSpeed = 290;
+        descriptionShort = CSTRING(HEAT_655_CS_Desc);
+        displaynameshort = CSTRING(HEAT_655_CS_Short);
+        ammo = "Cav_m3maaws_HEAT_CS";
+        type = "2* 256";
+        mass = 72.769997;
+        ace_overpressure_priority = 99;
+        ace_overpressure_angle= 50;
+        ace_overpressure_range = 2;
+        ace_overpressure_damage = 0.69999999;
+    };
+};
