+++ addons/weapons/config.cpp
@@ -3,14 +3,15 @@
 class CfgPatches {
     class ADDON {
         name = COMPONENT_NAME;
-        units[] = {};
+        units[] = {"Cav_m3maaws_HEAT_CS"};
         weapons[] = {};
         requiredVersion = REQUIRED_VERSION;
         requiredAddons[] = {
             "cav_main",
             "cav_common",
 
-            "tf47_m3maaws"
+            "tf47_m3maaws",
+            "cav_fix_tf47_m3maaws"
         };
         author = ECSTRING(main,CavModTeam);
         authors[] = {"Grey.V", "Brostrom.A (Evul)"};
@@ -20,4 +21,4 @@ class CfgPatches {
 };
 
 #include "CfgAmmo.hpp"
-#include "CfgCloudlets.hpp"
+#include "CfgMagazines.hpp"
