+++ addons/weapons/script_component.hpp
@@ -1,0 +1,5 @@
+#define COMPONENT weapons
+#define COMPONENT_BEAUTIFIED Weapons
+
+#include "\z\cav\addons\main\script_mod.hpp"
+#include "\z\cav\addons\main\script_macros.hpp"
