+++ addons/weapons/stringtable.xml
@@ -1,0 +1,13 @@
+<Project name="Cav">
+    <Package name="Weapons">
+        <Key ID="STR_Cav_Weapons_HEAT_655_CS">
+            <English>HEAT 655 CS</English>
+        </Key>
+        <Key ID="STR_Cav_Weapons_HEAT_655_CS_Short">
+            <English>HEAT 655 CS</English>
+        </Key>
+        <Key ID="STR_Cav_Weapons_HEAT_655_CS_Desc">
+            <English>High Explosive Anti tank Confined Space</English>
+        </Key>
+    </Package>
+</Project>
